feat: implement input processing strategy for Tana conversion
- Introduced a strategy pattern for processing different input types in `input-processors.ts`, enhancing modularity and maintainability.
- Created a factory in `processor-factory.ts` to select the appropriate input processor based on content type.
- Refactored the `convertToTana` function in `index.ts` to utilize the new processing strategies, improving code clarity and separation of concerns.
- Added comprehensive documentation for new classes and methods to ensure clarity on usage and functionality.
